+++ test/core/database/dao/collection_dao_test.dart
@@ -53,6 +53,7 @@ void main() {
   late MockTvShowDao mockTvShowDao;
   late MockVisualNovelDao mockVisualNovelDao;
   late MockMangaDao mockMangaDao;
+  late MockAnimeDao mockAnimeDao;
   late MockCustomMediaDao mockCustomMediaDao;
   late CollectionDao dao;
 
@@ -64,6 +65,7 @@ void main() {
     mockTvShowDao = MockTvShowDao();
     mockVisualNovelDao = MockVisualNovelDao();
     mockMangaDao = MockMangaDao();
+    mockAnimeDao = MockAnimeDao();
     mockCustomMediaDao = MockCustomMediaDao();
     dao = CollectionDao(
       () async => mockDb,
@@ -71,6 +73,7 @@ void main() {
       movieDao: mockMovieDao,
       tvShowDao: mockTvShowDao,
       visualNovelDao: mockVisualNovelDao,
+      animeDao: mockAnimeDao,
       mangaDao: mockMangaDao,
       customMediaDao: mockCustomMediaDao,
     );