+++ test/helpers/mocks.dart
@@ -21,6 +21,7 @@ import 'package:xerabora/core/database/dao/game_dao.dart';
 import 'package:xerabora/core/database/dao/movie_dao.dart';
 import 'package:xerabora/core/database/dao/tv_show_dao.dart';
 import 'package:xerabora/core/database/dao/custom_media_dao.dart';
+import 'package:xerabora/core/database/dao/anime_dao.dart';
 import 'package:xerabora/core/database/dao/manga_dao.dart';
 import 'package:xerabora/core/database/dao/visual_novel_dao.dart';
 import 'package:xerabora/core/database/dao/tag_dao.dart';
@@ -98,6 +99,8 @@ class MockTvShowDao extends Mock implements TvShowDao {}
 class MockVisualNovelDao extends Mock implements VisualNovelDao {}
 
 class MockMangaDao extends Mock implements MangaDao {}
+
+class MockAnimeDao extends Mock implements AnimeDao {}
 
 class MockCustomMediaDao extends Mock implements CustomMediaDao {}